src/api/requirements.txt
fastapi
fastapi[standard]
requests
gunicorn==23.0.0
azure-search-documents==11.6.0
azure-keyvault-secrets
aiohttp~=3.11.16
python-dotenv~=1.1.0
azure-cognitiveservices-speech~=1.43.0
jsonlines
nbconvert
pandas
azure-monitor-opentelemetry-exporter~=1.0.0b37
opentelemetry-instrumentation-fastapi~=0.54b0
jupyter
opentelemetry-instrumentation~=0.54b0
azure-identity==1.22.0
azure-ai-projects==2.0.0b3
gql==3.5.3
openai>=2.8.0
